Usunalem using namespace + kilka poprawek
