+++ text_analysis.h
@@ -3,29 +3,7 @@
 #include<string>
 #include<vector>
 //Slowniczek slow kluczowych pomocnych przy szukaniu pytan
-const std::vector<std::string> dictionary = {
-	"Czy","czy",
-	"Ile","ile",
-	"Gdzie","gdzie",
-	"Jak","jak",
-	"Dlaczego","dlaczego",
-	"Po co","po co",
-	"Czemu","czemu",
-	"Z kim","z kim",
-	"Z czym","z czym",
-	"Na czym","na czym",
-	"Kiedy","kiedy",
-	"Kim","kim",
-	"W którym","w którym","W ktorym","w ktorym",
-	"Kogo","kogo",
-	"Za co","za co",
-	"Co","co",
-	"U kogo","u kogo",
-	"Jakiego","jakiego",
-	"Dla jakiego","dla jakiego",
-	"Dla kogo","dla kogo",
-	"Komu","komu"
-};
+extern const std::vector<std::string> dictionary;
 /* Typ enumeracyjny uzywany przy markerach*/
 enum TextType
 {
@@ -41,7 +19,7 @@ class Markers
 	TextType type;
 	int percentage_chance;
 public:
-	Markers(int s, int e, textType t, int percent) :start(s), end(e), type(t), percentage_chance(percent) {};
+	Markers(int s, int e, TextType t, int percent) :start(s), end(e), type(t), percentage_chance(percent) {};
 };
 std::vector<Markers> findQA(const std::string& text);
 #endif
