+++ text_bounding.h
@@ -5,12 +5,11 @@
 #include "opencv2/imgproc/imgproc.hpp"
 #include "opencv2/highgui/highgui.hpp"
 #include "opencv2/opencv.hpp"
-using namespace std;
 using namespace cv;
 //Klasa zawieraj¹ca (narazie) funkcjê, która buduje obramowanie wokó³ bloków tekstu. 
 class ImageProcess
 {
 public:
-	vector <Rect> findRectangles(Mat *img);
+	std::vector <Rect> findRectangles(Mat *img);
 };
 #endif
